requirements.txt
comtypes==1.1.10
pip==21.3.1
pywin32==302
pywinauto==0.6.8
setuptools==57.4.0
six==1.16.0
aes-cipher
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
et-xmlfile==1.1.0
idna==3.3
iniconfig==1.1.1
openpyxl==3.0.9
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest~=7.1.3
pytest-dependency==0.5.1
pytest-html==3.1.1
pytest-metadata==1.11.0
requests==2.26.0
selenium~=4.5.0
setuptools==57.4.0
toml==0.10.2
urllib3==1.26.6
pywin32==302
pywinauto==0.6.8
mss==6.1.0
numpy==1.21.4
Pillow==8.4.0
pandas==1.3.5
pycryptodome==3.12.0
pytest-xdist
parameterized
xlsxwriter
pytest-bdd~=6.0.1
python-dateutil~=2.8.2
paramiko~=2.12.0
pytz~=2022.7.1